Fix Bluetooth phone audio on secondary users
This fix causes the BluetoothInCallService to be enabled for the system
user when needed, not just for the secondary user involved.

This change was based on a comparison of logs and on the observation
that when making a phone call from the main user using Bluetooth, there
is a period of time in which the proper behavior is exhibited in the
secondary user, because the system's BluetoothInCallService component
has yet to be disabled by AdapterService.

Issue: calyxos#1130
Change-Id: I310829ce71e8a6c7255b7b99f66aea3fea7d20d3